Adiciona os testes da camada service de evento

+++ src/main/java/com/eventoapp/services/EventoService.java
@@ -3,7 +3,6 @@ package com.eventoapp.services;
 import com.eventoapp.models.Evento;
 import com.eventoapp.repository.EventoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
@@ -23,6 +22,7 @@ public class EventoService {
     }
 
     public void salvarEvento(Evento evento) {
+        //TODO: Verificar se há um evento igual (nome e local) no banco
         eventoRepository.save(evento);
     }
 
@@ -33,12 +33,26 @@ public class EventoService {
     }
 
     public boolean deletarEvento(Long codigo) {
-        try {
+        Evento retorno = this.eventoId(codigo);
+
+        if (retorno != null) {
             eventoRepository.deleteById(codigo);
             return true;
-        } catch (EmptyResultDataAccessException error) {
+        } else {
             return false;
         }
+    }
+
+    public void atualizarEvento(Evento eventoAntigo, Evento eventoAtualizado) {
+        //TODO: Verificar se o evento foi alterado (ambos são iguais)
+        //Se os eventos não foram alterados, não salvar no banco
+
+        eventoAntigo.setNome(eventoAtualizado.getNome());
+        eventoAntigo.setLocal(eventoAtualizado.getLocal());
+        eventoAntigo.setData(eventoAtualizado.getData());
+        eventoAntigo.setHorario(eventoAtualizado.getHorario());
+
+        this.salvarEvento(eventoAntigo);
     }
 
     public void mensagemSucesso(RedirectAttributes attributes, String messagem) {
